hvdc_pipeline/requirements.txt
# HVDC Pipeline Requirements
# HVDC 파이프라인 의존성

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Configuration Management
PyYAML>=6.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning & Anomaly Detection
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
pathlib2>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Logging
loguru>=0.7.0
